Notification class
